gst-plugin-mle: use correct datatype for logging variables

+++ gst-plugin-mle/mle_engine/ml_engine_intf.h
@@ -238,7 +238,8 @@ public:
 
   ~Timer () {
     uint64_t end = GetMicroSeconds();
-    MLE_LOGD("%s: %llu us", str.c_str(), end - begin);
+    MLE_LOGD("%s: %llu us", str.c_str(),
+        static_cast<long long unsigned>(end - begin));
   }
 
   uint64_t GetMicroSeconds()
